Promote NextActionDataParser to top level
